microservice-refactor-agent/requirements.txt
ast
gitpython>=3.1.0
pygit2>=1.12.0
astroid>=2.15.0
pylint>=2.17.0
black>=23.0.0
isort>=5.12.0
click>=8.1.0
pydantic>=2.0.0
networkx>=3.0
matplotlib>=3.7.0
jinja2>=3.1.0
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0